requirements.txt
livekit-agents[codecs]>=1.4.6
livekit-plugins-openai>=1.4.6
livekit-plugins-sarvam>=1.4.6
python-dotenv>=1.0.0
aiohttp>=3.10.0
 